Move config dir functions to public functions in uv_fs::config
